First commit -- partway to modularity
I'm starting with the base files from my PhD thesis, and putting
segments that multiple files share into includes. As I update the files
to use the new includes, I'll add them to the repository.

I've added a Makefile to render the files, and the font that the axes
render requires -- which *should* be one I'm allowed to include in a
public repo, but I'll have to double-check that.